requirements.txt
amqp==2.3.2
billiard==3.5.0.3
cairocffi==0.8.1
CairoSVG==2.1.3
celery==4.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cssselect2==0.2.1
defusedxml==0.5.0
Django==2.0.6
django-paypal==0.5.0
django-rosetta==0.8.1
html5lib==1.0.1
idna==2.6
kombu==4.2.1
microsofttranslator==0.8
mysqlclient==1.3.12
pdfrw==0.4
Pillow==5.1.0
polib==1.1.0
pycparser==2.18
Pyphen==0.9.4
pytz==2018.4
redis==2.10.6
requests==2.18.4
six==1.11.0
tinycss2==0.6.1
urllib3==1.22
vine==1.1.4
WeasyPrint==0.42.3
webencodings==0.5.1
